requirements.txt
flask-socketio
paho-mqtt